% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teste.R
\name{teste}
\alias{teste}
\title{teste}
\usage{
teste(g)
}
\arguments{
\item{g}{Função a ser testada}
}
\description{
Função para testar se o desenvolvimento dos exercicios propostos está ok
}
\examples{
soma = function(...){
x = list(...)
s = 0
for(i in 1:length(x)) {
 s = s + x[[i]]
}
return(s)
}

teste(soma)
}
